Added new excercises
I integrated the latest excercises that use the API from Timothy Cannon. However, the VIs he wrote are a strict dependency. They would ideally be built as a separate package and managed as a dependnecy. I want to fix this long term, but this is so we can use it for an event today at UT